refactor(contribs): restructuration plus logique de GetContribs et FormatContribs
